Renamed or_interactivemarker to or_rviz.
