+++ include/or_rviz/markers/JointMarker.h
@@ -34,7 +34,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <openrave/openrave.h>
 #include <interactive_markers/interactive_marker_server.h>
 
-namespace or_interactivemarker {
+namespace or_rviz {
 namespace markers {
 
 class JointMarker;
