+++ plugins.xml
@@ -1,5 +1,5 @@
-<library path="lib/or_interactivemarker_rviz">
-    <class type="or_interactivemarker::rviz::EnvironmentDisplay"
+<library path="lib/or_rviz_rviz">
+    <class type="or_rviz::rviz::EnvironmentDisplay"
            base_class_type="rviz::Display">
         <description>
             Display information about an OpenRAVE environment.
